# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  handlers_auth.recoveryRequest:
    properties:
      login:
        type: string
    type: object
...
  model.Employee:
    properties:
      education:
        type: string
      email:
        type: string
      experience:
        type: integer
      firstName:
        type: string
      gender:
        type: string
      phone:
        type: string
      role:
        type: integer
      secondName:
        type: string
      specializations:
        items:
          type: integer
        type: array
      surname:
        type: string
      user_id:
        type: integer
    type: object
...
  model.Patient:
    properties:
      birthDate:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      password:
        type: string
      phone:
        type: string
      secondName:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
  model.PatientWithoutPassword:
    properties:
      birthDate:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      phone:
        type: string
      secondName:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
...
  model.ScheduleTable:
    properties:
      dates:
        description: '["01.06.2025", "02.06.2025", ...]'
        items:
          type: string
        type: array
      table:
        additionalProperties:
          additionalProperties:
            $ref: '#/definitions/model.UpcomingDoctorAppointment'
          type: object
        description: table[date][time] = запись или nil
        type: object
      times:
        description: '["09:00", "09:30", "10:00", ...]'
        items:
          type: string
        type: array
    type: object
...
  model.TodayAppointment:
    properties:
      date:
        type: string
      id:
        type: integer
      patient:
        type: string
      patient_id:
        type: integer
      time:
        type: string
    type: object
...
  model.UpcomingDoctorAppointment:
    properties:
      id:
        type: integer
      patient:
        type: string
      patient_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: ПО для стоматологической клиники
  title: Примастом API-Gateway
  version: "1.0"
paths:
...
  /api/appointments-today:
    get:
      description: Возвращает список записей на сегодня для авторизованного врача
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TodayAppointment'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiCookieAuth: []
      summary: Получить записи на сегодня
      tags:
      - Врач
...
  /api/employee-password-recovery:
    post:
      consumes:
      - application/json
      parameters:
      - description: Логин
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers_auth.recoveryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Восстановление пароля сотрудника
      tags:
      - Авторизация
  /api/employee-registe:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные сотрудника
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Employee'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Регистрация сотрудника
      tags:
      - Администратор
...
  /api/password-recovery:
    post:
      consumes:
      - application/json
      parameters:
      - description: Логин
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers_auth.recoveryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Восстановление пароля пациента
      tags:
      - Авторизация
...
  /api/patient/me:
    get:
      description: Возвращает информацию о текущем авторизованном пациенте на основе
        cookie access_token
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PatientWithoutPassword'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiCookieAuth: []
      summary: Получить данные пациента
      tags:
      - patient
...
  /api/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пациента
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Patient'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Регистрация пациента
      tags:
      - Администратор
  /api/register-in-clinic:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пациента
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.PatientWithoutPassword'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Регистрация пациента в клинике (без пароля)
      tags:
      - Администратор
...
  /api/schedule-with-appointments:
    get:
      description: Возвращает таблицу расписания будущих записей для врача
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ScheduleTable'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiCookieAuth: []
      summary: Получить будущие записи
      tags:
      - Врач
...
swagger: "2.0"